add host interaction weight to the hostdb
