+++ modules/renter/hostdb/consts.go
@@ -7,17 +7,13 @@ import (
 )
 
 const (
-	// maxHostDowntime specifies the maximum amount of time that a host is
-	// allowed to be offline while still being in the hostdb.
-	maxHostDowntime = 10 * 24 * time.Hour
+	// historicInteractionDecay defines the decay of the HistoricSuccessfulInteractions
+	// and HistoricFailedInteractions after every block for a host entry.
+	historicInteractionDecay = 0.9995
 
-	// minScans specifies the number of scans that a host should have before the
-	// scans start getting compressed.
-	minScans = 12
-
-	// maxSettingsLen indicates how long in bytes the host settings field is
-	// allowed to be before being ignored as a DoS attempt.
-	maxSettingsLen = 10e3
+	// historicInteractionDecalLimit defines the number of historic
+	// interactions required before decay is applied.
+	historicInteractionDecayLimit = 500
 
 	// hostRequestTimeout indicates how long a host has to respond to a dial.
 	hostRequestTimeout = 2 * time.Minute
@@ -26,13 +22,30 @@ const (
 	// scan.
 	hostScanDeadline = 4 * time.Minute
 
+	// maxHostDowntime specifies the maximum amount of time that a host is
+	// allowed to be offline while still being in the hostdb.
+	maxHostDowntime = 10 * 24 * time.Hour
+
+	// maxSettingsLen indicates how long in bytes the host settings field is
+	// allowed to be before being ignored as a DoS attempt.
+	maxSettingsLen = 10e3
+
+	// minScans specifies the number of scans that a host should have before the
+	// scans start getting compressed.
+	minScans = 12
+
+	// recentInteractionWeightLimit caps the number of recent interactions as a
+	// percentage of the historic interactions, to be certain that a large
+	// amount of activity in a short period of time does not overwhelm the
+	// score for a host.
+	//
+	// Non-stop heavy interactions for half a day can result in gaining more
+	// than half the total weight at this limit.
+	recentInteractionWeightLimit = 0.01
+
 	// saveFrequency defines how frequently the hostdb will save to disk. Hostdb
 	// will also save immediately prior to shutdown.
 	saveFrequency = 2 * time.Minute
-
-	// historicInteractionDecay defines the decay of the HistoricSuccessfulInteractions
-	// and HistoricFailedInteractions after every block
-	historicInteractionDecay = 0.999
 )
 
 var (
